bootloader: waf: define `_THREAD_SAFE=1` when compiling on/for AIX
When compiling bootloader for AIX, add `_THREAD_SAFE=1` to the
list of defines to be passed to compiler. This supplements the
`_REENTRANT` define that we seem to be using in-lieu of `-pthread`
compiler flag; while on most platforms, `-pthread` defines
`_REENTRANT`, on AIX, it defines `_THREAD_SAFE`.

`_THREAD_SAFE` is also defined (if necessary) when `pthread.h`
header is included on AIX, but if we were to rely on this, we
would need to ensure  it is the first header included in every
compilation unit.

Having `_THREAD_SAFE` (or `_THREAD_SAFE_ERRNO`) defined enables
thread-safe `errno`. In the context of bootloader, this might
matter only when splash screen is enabled (since that runs in
a secondary thread), so this change is done mostly for the sake
of consistency (since we also pass `-lpthread` to the linker).